Fix the lexer to use the old name model.
The new naming model that allowed punctuation in alphabetic names
really wasn't doing us any favors.  This commit rewinds the design
to match the C# implementation, more-or-less, with the small caveat
that - must be followed by an alphabetic character.  That caveat
allows 'x-1' to be lexed as 'x' '-' '1' instead of as a compound,
which was a defect of the lexer grammar used in the C# version.