Shuffle the opcodes again; add special ops for arithmetic.
In an attempt to get Smile to run as fast as possible, certain binary/unary forms will now be recognized by the compiler and emitted as special opcodes that will short-circuit the usual call chain if they recognize the data type.  This is mainly so that arithmetic things like 1+2 don't involve a giant pile of heap allocations and function calls.