Restructure compiler to use intermediate code.
Compiler previously emitted the bytecode directly, and every instruction
was responsible for fixing its own branch addressses.  This is unworkable
for more complex scenarios, so this begins a major restructuring to emit
intermediate instructions as a tree of basic blocks with the instructions
set up as a linked list in each block so that code motion is simple, free,
and easy.  Address resolution (and branch target resolution) is done using
subsequent passes over the data, so that the final compiled bytecode is
simpler and cleaner.

This commit makes the compiler *mostly* correct for most of the core forms,
but till-loops are still not done.  None of this is tested yet either.