Fix nasty multi-compile bug by binding Segments to their Tables.
In scenarios where you compile functions separately, their bytecodes may reference CompiledTables sets that are unrelated.  For example if 'f' is compiled, and then 'g' is compiled but it calls 'f', then during 'g' calling 'f', 'f' may end up referencing objects or strings in the CompiledTables for 'g'.  This was bad, and led to segfaults.

So the fix is threefold:  First, to have a compiledTables field on every ByteCodeSegment that points to its data set.  Second, to update everything to stop referencing CompiledTables directly.  And third, to make Eval-and-friends switch _compiledTables any time it switches _segments.

This does come with a few nice side benefits: The return value of the various Compile() functions now returns a UserFunctionInfo (a function) instead of a Compiler, which makes considerably more sense.  And Eval_Run() now just takes in a parameterless UserFunctionInfo to invoke, which is a *lot* better.

Still, it's a pretty substantial change, despite its conceptual simplicity, and it touches a lot of code and has some small performance implications for function calls.  But it's far more correct, and works in a much wider range of scenarios, so we're going with it for now.