Extend the binary line-break rule to apply between unary ops and their operands.
There is no scenario where line-breaking between a unary operator and its operand is a good idea in real-world code, so this actively prohibits it, which ensures that line-breaks are more likely to result in syntax errors when the programmer expects them to --- at the end of bad expressions.

Consider this (valid) example:

    if x < -y then x = -y

If we break lines before the last '-', it becomes:

    if x < -y then x = -
        y

This makes it too easy for the subsequent 'y' to get lost.  Consider what happens if you were to allow the breaks and then write this code:

    if x < -y then x = -
    [Stdout.print x]

Without this rule, the parser would treat that example as acceptable, which is unlikely to be what anybody wants.  With this rule, the parser will treat the above as an error, which is almost always going to be the correct answer.

This is a useful safety check, and one that prevents bugs in real code.  As usual, you can always override the default parsing behavior with parentheses, so that if you really need to write "-y" on two lines, you can always write it as

    (-
      y)

Even though Smile is generally free-formatted, most programmers will prefer errors to appear for common mistakes, and this rule ensures many more errors will get caught at the minor inconvenience of occasionally requiring more parentheses around multiline expressions.