Add new built-in special '$dot' symbol.
Lists of the form [$dot x y] will soon replace the current (x.y) pair
construct, so the first thing to do is add '$dot' as a known special symbol.

As of this point, all unit tests pass, but it's unlikely that they will
pass after the next commit until most of this refactor is complete.