Refactor #include to be a true module system.
This might yet have bugs, but the simple test case of double-inclusion
(a uses exports from b, b uses exports from c) now works as expected, at
least for variables.

Exported #syntax declarations are not yet imported into the caller, so
this still isn't quite a complete implementation, but it's finally enough
for a few simple test cases to work, and it correctly respects scoping
rules and whether a particular variable was locally-declared or imported.